internal/command/auth: add support to configure docker auth directly
Currently `flyctl auth docker` has a depenedency on having Docker
installed. However, there are many tools that reuse the Docker config
for OCI management. This commit adds support to update the Docker
config.json directly when the `docker` lookup fails.

It tries to be lazy about evaluating the different parts of the JSON
file as to not accidentally drop other config entries.

Signed-off-by: Maisem Ali <maisem@tailscale.com>