2/2: Better support for dockerignored Dockerfiles
A dockerignore file has rules that prevent files from being
sent to the Docker daemon as part of the build context and/or
being included in the built image. However the Docker daemon
itself also requires the Dockerfile to be contained inside the
build context. Therefore tools like flyctl and docker-cli have
to have special case handling of the dockerignore file.

When the dockerignore file contains a rule that matches the
Dockerfile or the dockerignore file itself, the build context
must nonetheless include both files, and the container engine
will handle it correctly.

The existing implementation works when the file is named "Dockerfile",
but not when the user manually specifies a different path (such as using
--dockerfile or using build.dockerfile in fly.toml), or when
the Dockerfile is a symbolic link. These cases would result in
an error on fly deploy with file not found from the container engine.