Add dockerignore generation to `fly launch`
If a .dockerignore file does not exist and .gitignore files do exist, generate a single .dockerignore file that combines all the .gitignore files. The .dockerignore file is created in the same directory as fly.toml.

If no .gitignore files are found and no .dockerfile exists, `fly launch` will print a message indicating a .dockerignore file should be created.