implment a new flag: --ignorefile string
  Path to a Docker ignore file. Defaults to the .dockerignore file in the working directory.

Rationale: the hope/plan is to support a path where scanning and dockerfile generation is
done at every deploy time rather than only once at launch time.  This has two benefits:
it adjusts the Dockerfile as needs change (new version of ruby, addition of a gem that
requires a debian library to be installed, etc); and may be less scary to the Dockerphobic.

If we are going to make the Dockerfile invisible, it makes sense to make the .dockerignore
file invisible too; it will be generated by the same scan process, which will likely be
adjusted to use .gitignore as a starting point.