fly deploy for machines for apps v2
We're pushing to migrate the Fly Apps platform to machines instead of nomad to improve developer experience. The first big step will be make new apps use the machines platform instead of nomad. There's a bunch of steps to get there, and this is one of them...

Update `fly deploy` when we're working on an app with machines to:

1) Prompt developer to migrate the machines into the "Fly Apps platform". The machines being deployed will become the instances of the apps. Specifically, wWe put some metadata on those machines indicating they are part of the Fly Apps platform. Subsequent deploys will only operate on machines with that metadata.

2) We create releases in the graphql api for each deployment. The release only gets created and marked as pending for now. It does not get updated. That's ok for now.