fly m restart: remove double lookup for machine IDs
We checked to make sure the machines could be found at all, then checked to make
sure they shared the same app. this was nice, but it hit the API a *lot* more than it should.

Instead, we just look up the machines from within the app specified by the first machine,
and if the machine is missing, gently suggest that it may not be in the same app.

This way, the UX doesn't regress too far (we still nudge the user in the right direction),
but we also don't spam the API for little reason.