Improve long help message for `secrets list`
The long message appears broken on the terminal. This CL not only
removes the broken indentation but also fixes its wording and amends its
alignment.

Before:

```
$ flyctl help secrets list
List the secrets available to the application. It shows each
                secret's name, a digest of the its value and the time the secret was last set.
                The actual value of the secret is only available to the application.

[snip]
```

After:

```
$ ./bin/flyctl help secrets list
List the secrets available to the application. It shows each secret's
name, a digest of its value and the time the secret was last set. The
actual value of the secret is only available to the application.

[snip]
```