`fly apps restart`: begin supporting Appsv2 properly
This is a work in progress.
It should only restart the machines that are orchestrated by Apps.
This has had *some* testing but not rigorous enough for release.

Additionally, it should create a release in the app,
visible through the fly.io web UI or `fly apps releases`.
This does not work, but it's unclear whether this is a flyctl
issue or a backend issue right now.
The code that does this is a little ugly anyway, so this will likely
need a little bit of polish before it's ready to be released :)