WIP: update fly config save for apps v2
update the gql schema, add a gql query to get the current release, and use the config definition from that release to build out the config that gets written.

This does not work yet; definition is type interface{}, but we need somethign like map[string]interface{} to work with existing appv2.FromDefinition(). alternatively we could add something like appv2.FromInterface()