`fly machine update`: fix hanging bug
two bugs combined to make something pretty nasty

`internal/machine/update.go` created a synthetic machine object
(`&api.Machine{ID: input.ID}`) which didn't have an `instanceId`

This usually wasn't an issue, but when `WaitForStartOrStop` was waiting for
`stop` (this is an edge case), the API call expected to have an instance ID,
and so it returned an error.

Just so happens that our error handling in `internal/machine/wait.go` didn't
recognize this specific error code, and just assumed we needed to wait longer.

oops!