first pass at fixing `fly launch` not saving the chosen region
The primary_region should be saved in `fly.toml` for v2 apps.
this isn't part of the V1 `fly.toml` file, so there's some weird
special casing going on here.

todo: does this affect `fly config save`? should it?
does this break anything else?
(it's possible that other places that read fly.toml won't like
that it's now perfectly OK for it to have `primary_region`.
granted, I think it was supposed to be that way already, but you
can never be too safe!)