Remove confusing msg from fly deploy
Remove the "Migrate %d existing machines into Fly Apps Platform?" confirmation from `fly deploy`.

Previously developers got that message if they were deploying to machines that that were missing the apps v2 metadata. This has just been confusing for developers with little upside.

The original idea was it would give folks a warning if they were deploying and had a heterogenous cluster that they didn't want to trample over. The reality is that `fly deploy` has always applied to all machines, so this confirmation was just confusing.

I debated adding in a check for different images/other configs that would be material, but decided it wasn't worth it. Most developers will use `fly deploy` to update all machines in their app, so we should just do that for them without prompting. Developers that need to update a subset of machines will likely be using `fly m update` or the machines api already anyway (and those are the right answer).