Fix Context error handling in WaitForStartOrStop
Under the hood, the flaps client is using the
github.com/PuerkitoBio/rehttp package for HTTP transport. When a request
is canceled by a Context cancellation or timeout, this package returns
an error that does not wrap context.Canceled or
context.DeadlineExceeded. As a result, WaitForStartOrStop does not
handle these cases properly. These errors are treated like any other
error: flyctl will wait for a backoff period and then try again, ad
infinitum.

To get around this, we can just examine the error associated with
waitCtx.