Allow creation of flaps clients with only an app name
When we use api.machines.dev, we need only the app name (not an
AppCompact) to create a flaps client. Many flaps client users get an
AppCompact from the API only to create the client, so we can often save
a round trip by requiring only an app name to create one. NewFromAppName
provides this. (When access over Wireguard is requested, the AppCompact
is still needed, so it will automatically fetch it when this happens.)