Don't allocate PTYs when a command is specified
This makes `fly ssh console` work more like OpenSSH does by default. If a
command is specified with `-C`, a pseudo-terminal will not be allocated unless
`--pty` is used. (Though it would be nice to use `-t` to harmonize with `ssh`,
it's already taken in `flyctl`.)

As an exception, `sh`, `/bin/sh`,`bash`, and `/bin/bash` with no further
arguments do get PTYs for the time being, in case users are used to this just
working. A warning is printed to let users know that the default behavior is
going to change.