Make 'fly machine clone --attach-volume' error if source doesn't have volumes
Instead of silently ignoring the flag, return an error clarifying that
the source machine must have a volume.

This also leaves an opportunity to change the flag arguments in the
backward compatible way later and allow users to specify the mointing:

  --attach-volume <volume-id>[:<mount point>]

In this form we can clone the machine and attach the volume even if
the source doesn't have any volumes.