fly consul attach|detach
Add a new command for attaching and detaching a Consul cluster to an app. We borrow the postgres launcher functionality for provisioning a Consul cluster, and set FLY_CONSUL_URL with the cluster url.

Detach will unset the secret with the cluster url.

This works for any app. Going forward the `fly consul attach` command should be used instead of the fly.toml experimental `enable_consul` configuration field.