Move lock files to ~/.fly to fix multi-user flyctl
Right now, when running flyctl in a multi-user environment, you run into
permission issues with the shared lock files in /tmp:

* `flyctl.cache.lock`
* `flyctl.config.lock`
* `flyctl.agent.lock`
* `flyctl.agent.start.lock`

So, I've moved them all to `~/.fly`. In `~/.fly` we already have
`config.yml` and `fly-agent.sock`, so it makes sense that the lock files
are also not shared.

Closes #977