Perform sanity checks after Apps V2 deploy
Now that the healthchecks are disabled by default, it's very easy
to rollout a broken version on every machine in the app.

This PR introduces a "sanity check" step during the rollout. We are
gonna wait for up to 10 seconds to ensure that the newly deployed
machine is up and running and not restarting with non-zero exit code.
This won't catch all errors, but should prevent the most obvious ones
like OOMs right after deploy, missing configuration, etc.