Add `Dockerfile` to build context regardless of path
We currently have an override to ensure that `/Dockerfile` and `/dockerfile`
are added to the build context even if they're excluded by the `.dockerignore`.
However, if the `Dockerfile` is located elsewhere, we should add an override
for *that* path instead.

Thanks to jacobgreenleaf for originally working on this in PR 1213. These
changes are similar.