Ignore HTTP_PROXY for tunneled remote builder connections
If host-wide HTTP_PROXY is set, `flyctl` fails to connect to the
remote builder because Docker client tries to connect to the daemon
via proxy.
We can't really proxy request to remote builder, as the proxy knows
nothing about the tunnel established by `fly agent`.

So, use a custom HTTP client for Docker that ignores proxy settings
from environment.