No TOML `omitempty` for `AutoStartMachines`/`AutoStopMachines`
The encoder already omits these fields if `nil`, since `nil` can't be
represented in TOML. The `omitempty` actually causes them to be omitted if
non-`nil` and `false` as well. This is particularly bad for
`AutoStartMachines`. If `fly config save` pulls down `false`, then the field
will be omitted from the written `fly.toml`. If `fly deploy` subsequently uses
the `fly.toml`, then it gets set to `nil` (which defaults to `true`)!