Remove local test parallelism
I've been trying to debug why this doesn't work for a while, and this is me (kind of) giving up :(. I'll dump some info here for myself (and anyone else debugging this) in the future.

So initially, I thought this was a [web issue](https://flyio.discourse.team/t/preflight-tests-failing-with-500-errors/3605), but after spending quite a bit of time with this, I can't be sure. It looks like *something* in fly can't handle mutliple commands runnning at the same time, possibly because some state is overwritten. For now, I think the best way of handling this is just disabling go test concurrency, whhile keeping it in Github Actions.