Ignore destroyed volumes
Now that we use the Machines API to work on volumes, the volume states
that we receive correspond directly to the volume states in `flyd`.
Several of these states are for volumes that have been effectively
"destroyed" but aren't truly destroyed yet. Filter out volumes in these
states so that we don't try to use them.

`GetAllVolumes` provides an unfiltered version.