Add Metrics Regarding Flaps Calls (#2545)
* Suggest count flag for vol on deploy errors

* `analytics`: revive back `flaps_call_metrics` branch

* Fix flaps call bug not seeing /[machine_id] calls

* Fix rebase

* Linting

* Use structs to store info on flaps actions

flaps actions now use a structured format when performing an action,
like getting machine data, refreshing a lease, etc. Some flaps actions
need to have some context regarding the path and what they're doing.
It's easier to have some structure regaridng what flaps calls are doing,
so we can extract metrics from that action, instead of parsing regex
paths and guessing based on POST data.

I'm writing this commit while I'm tired, sorry.

* Fix `fly console` command

For whatever reason, we were spawning a new context in when cleaning up
the ssh console machine, instead of using the existing one. There isn't
a documented reason, and as far as I can tell, this is the only place in
the codebase where we do this.

* Properly get status code from flaps

Doing this instead of guessing is a. the obvious correct choice and b. fixes the final broken preflight test (TestAppsV2MigrateToV2_Autoscaling)

* Linting

* Removed out of scope FIXME

* Send duration in flaps call metrics

* Don't send 0 status codes

* Allow any binary name when sending metrics

* Use url.Values + return statusCode in _sendRequest

* flapsAction to file + derive method from action

* Check for cmdCtx.Parent being nil in for loop

---------

Co-authored-by: William Batista <bootlegbilly@protonmail.ch>