Speed up `fly volumes` commands by switching to `GetAppBasic` (#3015)
Switch `fly volumes` commands to `GetAppBasic`

They're currently using `GetApp`, which reports a lot of information,
including a list of Machines. For apps with a lot of Machines (including
destroyed ones), this can take a few seconds for the API to prepare.

These commands don't need this information, so use `GetAppBasic`
instead. This speeds them up, and it may reduce a bit of load on the API
too!