Make Launch v2 the default 🎉 (#2956)
Switches the default to the new Launch interface.

This interface builds a launch "plan," then asks the user to confirm that plan. If they don't like the generated defaults, it sends them to a web form to finish up configuring their app. Once that's done, it continues launching on the CLI.

The previous Launch behavior is kept behind the hidden `--legacy` flag, but this is deprecated and will be removed soon (probably alongside the removal of the migrate-to-v2 tests, which depend on the legacy behavior)