Allow changing CLI configuration directory (#2785)
* Allow changing configuration directory

We should introduce an environment variable that, if set, will be used
instead of hardcoding `~/.fly` everywhere. This allows for the user to
use the CLI in non-standard environments, or in more advanced ways (for
example, by swapping the config dir to isolate separate accounts).

* Simplify syntax of `GetConfigDirectory`

* go fmt

* Remove test reference to `DetermineUserHomeDir`

This function was removed.

* Fix config dir path in tests on Windows

This was erroneously using `path` instead of `filepath`, which meant on
Windows the generated path was using the wrong slash direction.