Lay foundations for xdg base dirs
This commit prepares the codebase for a future implementation of
xdg base dirs. Until now, the whole codebase assumed there was one
"config dir" where anything flyctl-related went, even if it was
not a config file. This commit creates two new concepts: the "state
dir" and the "runtime dir", corresponding to XDG_STATE_HOME and
XDG_RUNTIME_DIR, respectively. However, the commit keeps the same
logic for determining the state dir and the runtime dir as the
config dir, so the behavior of flyctl should not be different in any
way.