Doctor: add dns and http tests (#3110)
* `doctor`: stub out WG DNS and HTTP tests

* `agent.Client.ConnectToTunnel()`: allow silent operation

* `agent`: add (untested) DNS txt lookup

* `doctor`: finish skeleton of DNS test

* `doctor`: add skeleton of WG HTTP check

* `doctor`: WG HTTP: use raw network connection for HTTP

The default go HTTP client didn't play nice with the agent-provided sockets for some reason. this works just as well, for this specific use-case.

* `doctor`: WG DNS: just resolve _api.internal

Turns out _peer.internal doesn't return interactive peers, which means we'd have to do a whole song-and-dance to get our test peer to show up in the list.

I don't like double-dipping on resolving flaps as the *only* wireguard testcase, but we can refine this over time.

* `doctor`: WireGuard: test sequentially.

Since we're testing on flaps dns resolution for the DNS test, the HTTP test will never succeed if the DNS test fails. Therefore, there's no need for the awkward indirect error handling now.

* `doctor`: WG HTTP: cleanup

* `doctor`: WG HTTP: remove unused hacky code