Allow 0 IPs when checking DNS + backoff
Firstly, I failed to kekep in mind in checkDNS that apps can have no
IPs. This commit makes it take into account. We also now run the query
repeatedly until we get a successful hit (which hopefully should just be
the first time we run it ;) ) up to a minute