Add a check for the fly's DNS address being ready (#3195)
* Add a check for the fly's DNS address being ready

This can help to mitigate issues where the new DNS address isn't
resolving yet (though it's not a fix).

* go mod tidy

* Add a flag to skip DNS checks

* Allow 0 IPs when checking DNS + backoff

Firstly, I failed to kekep in mind in checkDNS that apps can have no
IPs. This commit makes it take into account. We also now run the query
repeatedly until we get a successful hit (which hopefully should just be
the first time we run it ;) ) up to a minute

* go mod tidy

* Rename url to appURL

To avoid shadowing stdlib

* Use Quad9 (9.9.9.9) to check DNS

To stay consistent

* Change max backoff time to 30s