`launch`: `ha` warning: adapt to web changes
This took a bit of piping around, but basically, here's what the behavior *should* be now.

If you deploy to an org, and that org has no payment method, we should warn that you can't deploy HA. (this happened before)
If you *change* to an org that *does* have a payment method in the UI, you should then get the default HA or Not HA deployment that you would've gotten if you started on that org.

Likewise, if you start on an org with a payment method, and switch to one without, you should get the warning printed and have HA turned off.

This should also not double-warn.