`Flatten()`: move compute determination to own func
This allows launch to piggyback off the same logic. Initially I wanted to just copy-paste this into Launch and be done with it, but I had to copy `matchesGroups` too which just turns it into a big mess of unmaintainable code.

Not that this is *much* better, but it's a little better (I hope)