Cordon and gracefully shutdown old machines during blue-green deploy
If flyctl destroys the old machines immediatelly after uncordoning the
new ones, fly-proxy doesn't have any time to learn about the new machines
and has nowhere to send the requests.

So this patch modifies the procedure to first cordon the old machines.

Even with this in place, we need to gracefully shutdown the old machines
in order to let the app properly terminate existing connections by handling
SIGTERM.