Handle unwritable gem installation directories gracefully.
See: https://community.fly.io/t/rails-7-applications-do-not-need-a-dockerfile-rails-gem/18738/11

Background: while installing and running the dockerfile-rails gem has been working quite well
for some time, this fails in locked down environments such as a docker container for a Ruby
image when running as an unpriviledged user.

The purpose of dockerfile-rails (in addition to creating a Dockerfile in the first place) is
twofold:

 * Make adjustments to fly.toml to do things like allocate a volume for sqlite and run
   sidekiq in a process.
 * Be available make assisting people who have failed first deploys or make later
   changes.

While both of these are good reasons to install the gem when possible, neither of them
are sufficient reason to fail a fly launch if the installation won't succeed and
a Dockerfile is present; which generally will be the case for Rails 7.1 onward.

As to what to do should the gem installation directory be locked down and no
Dockerfile is present - should this come up, the recommendation will be to install
dockerfile-rails manually and run it before re-attempting fly launch.