For Rails, check Dockerfile for PostgreSQL (#3363)
Prior to Rails 7.1, Rails did not produce a Dockerfile, so we would
produce one, and set DATABASE_URL.

Now with Rails 7.1, a Dockerfile exists, and if it isn't set up for
PostgreSQL, the initial deploy will fail.  Overwriting the Dockerfile
would be a bit rude, so going forward we will only default to
PostgreSQL if we detect that the support is present.

Ths is a bit inconsistent with how we are handling MySQL, but the thought
being is that if you intentionally chose MySQL, that likely is what you
really want.