Fix default user/groups in buildpack builds.
The Pack library started supporting custom groups and users
owning files in the image - unfortunately the "default" behavior
changed to be "make all files owned by UID/GID 0" and the old
"make file ownership match the user running the process" behavior was
placed behind a magic constant for the new GroupID and UserID parameters
to the build arguments.

The symptom we saw is that files in buildpack-generated images were all
owned by root, while the running process was non-root, causing ENOACCES
errors when running images and VMs.

This fixes the problem by passing the new magic constant for GroupID
and UserID in buildargs for buildpack-based images.