Enable scanners to modify/append to error messages via a callback
This should help with the movement from launch SkipDeploy=true
where DeployDocs would suggest reading materials before deployment
to SkipDeploy=false and the link to the reading materials only
shows up on failure.

The callbacks are passsed the actual error message, so over time
it may even be possible to provide (or point to) specific advice
for common errors.