`fly console`: use all of `console` group's `[[vm]]` section
We currently check for the `[[vm]]` section that applies to the
`console` process group and then access the raw `memory_mb` field as it
is in `fly.toml`, which may be unset (causing a `nil`-pointer
dereference). Furthermore, we ignore the `memory` field, which allows
for a human-friendly suffix.

It seems best to just make `fly console` use *all* of the `[[vm]]`
section, including e.g. the `size`. If we feed that information to
`flag.GetMachineGuest`, then merging the `fly.toml` configuration
and any command-line overrides is very easy.

The minimum/maximum memory check/adjuster was written when `fly console`
accepted *only* flags to determine the Machine size, which explains the
checks for `--vm-memory` and `--vm-size` when deciding whether to
operate or whether to print anything. With the additional features it's
gained since then, I think it makes sense just to run it
unconditionally. (I still think that being able to type `--vm-cpus` and
have the memory automatically increased is kind of convenient, so I
decided not to remove it.)