Create machines for bluegreen deploy concurrently
Previously, we created the green (new) machines sequentially when running a bluegreen deployment. That works ok when there are a few machines, and it becomes a bit slow when there are 70 or 100 or 500+ machines in the app.

Update the CreateGreenMachines() func to use the pool pattern we have in a few places.

I did the same thing in the MarkGreenMachinesAsReadyForTraffic(), CordonBlueMachines(), StopBlueMachines() funcs.