Have fly launch stop parsing for arguments at a double dash (#3411)
The purpose of this is to pass the remainder of the arguments to
the language or framework specific scanner to do with as it wishes.
To start with, the framework scanners that include dockerfile
generators will pass the remaining arguments onto their respective
dockerfile generator.  Other scanners may find other uses.

Note that the chosen argument scanner, cobra, does not support this
at this time so the arguments are removed and retained before being
passed to cobra.