fix bug in `fly ssh console --user`
Previously, `fly ssh console --user username` would fail because the ssh certificate was always generated for the root user. This has been broken for some time, and became more prominent as we started using app specific certs in more places.

Pass through the `--user` flag to the certificate generation.

This also fixes the TestPostgres_ImportSuccess, which relies on the `--user` flag.