`fly vol snapshots ls`  works for deleted volumes
Previously, running `fly vol snapshots ls` would fail when provided the volume id of a deleted volume. We want to show snapshots from deleted volumes, because it enabled folks to create a new volume from those snapshots.

Update `fly vol snapshots ls` to work for destroyed volumes using the graphql api to retrieve the snapshots, instead of the flaps api. The one downside of doing it this way is the returned snapshots won't include status, but that seems like an ok tradeoff.