Skip deploy for common Rails dockerfile errors
Rails 7.1 will create a Dockerfile for new applications, and
after that you are on your own.

dockerfile-rails has been updated to check for common errors:

https://github.com/fly-apps/dockerfile-rails/commit/a4b81d26a5f657a5e125f71909611f0292a9eb96

When such are detected, it will return a special exit code of 42.

These changes to flyctl will cause the Rails scanner to set
SkipDeploy to true when such occurs, and to output instructions
on how to run `bin/rails generate dockerfile` to get an updated
Dockerfile.