Remove homebrew bump workflow (#3532)
Our homebrew version bump is failing because we're now in the autobump list. afaict this is a good thing.

Every 3 hours, homebrew runs an action to update all formulae in the autobump list. That's better than the daily update we were doing, yay!
https://github.com/Homebrew/homebrew-core/pull/168397
https://github.com/Homebrew/homebrew-core/blob/ef39119111a7a4786b66af3c4921cd05d5c642fc/.github/workflows/autobump.yml

We no longer need our own homebrew bump workflow. 🎉