Machine Based health Checks (#3430)
* Initial implementation of machine based tests

We basically take the command provided in [[services.machine_checks]],
and run it. If we get an exit code that isn't 0, we say that we failed
the health check. This is only implemented for the rolling strategy
right now, and needs to be implemented for bluegreen, canary, etc.

* Fix some machine check issues + add it to http_service

* Add test machine support for the canary strategy

We're also now parallelizing canary machine creation. This isn't worth
a fresh produce or anything since it was just a few lines of code, but
still neat

* Fix canary machines being destroyed

This was silly

* don't split command

otherwise, we need to put quotes around the command for it to work correctly

* Update fly-go and change process grouchap

* oops

* Set TEST_COMMAND instead of RELEASE_COMMAND

* Remove unnecessary comments

* Use %q instead of %s

* Add machine_checks to full-reference.toml

* Add test machines to machines_test

* Wait for logs using backoff

* added a test for machine checks

* Change TEST_COMMAND to FLY_TEST_COMMAND

* Add KillSignal/KillTimeout and switch Entrypoint/Command to string array

* Fix preflight

* Change KillSignal to string ptr

* Only test the services of a machine that match said machine's process group

* Check nil instead of len(svc.Entrypoint) > 0

* Add a test for ToTestMachineConfig with an existing machine

* Kill info changes

The comment says it all, but we basically should only use the app config's
kill signal/timeout if an image isn't set. Otherwise, we use the default
info for the image.

* oopsie

* Use the canary guest function

Honestly does exactly what we need

* Added a check for canary deploys

* Clean up some logic + add more test cases

* Fix crash when encountering errors

* Fix a test testing the wrong thing

* Fix NoImage + KillInfo + OrigMachine Kill Info test