Show warning for active incidents
This change adds a background task that queries status.flyio.net
for unresolved incidents for every command.

If any incident is returned, it will show a warning to the user.
The warning includes instructions to either check statuspage or
execute the specific command for listing the incidents which provides
extended information.