Add flyctl build step to preflight GPU tests
This duplicates effort a bit and might be worth investigating, but without it, flyctl isn't available to download when the preflight_gpu workflow runs on schedule.

Reusing the workflow directly would trigger an additional preflight (non-GPU) run. This may be OK--it's just a question of what scope we want for this change and whether or not extra preflight runs are worth less duplication.