Send buildID + build type to web
This PR consists of two changes:
Firstly, we're now setting the buildID as part of the build metadata.
That way, we can use that info in ui-ex to show some info about the
build related to a specific release.

Secondly, we're adding a new build type for depot, since it differs from
our current remote builds.