add fly deploy test for deploying with new mounts
The FlyDeployHA test already tests deploying with new mounts, but only
runs when the test environment is configured to use multiple regions
(at time of writing the CI pipeline only uses one primary region).
This commit adds a test of adding mounts through deploys that will run
when FlyDeployHA is skipped (and which will be skipped itself if it is
redundant with FlyDeployHA).