Generates a dedicated token for synthetics
To enable opening the synthetics system for Orgs to check their oen orgs, we'd need to know the orgs that a given flyctl agent has access.
This is a first step where we create a dedicated orgs read-only macaroon that enables the synthetics backend app to list all the orgs the flyctl agent has access, and use the result to register that agent instance to probe the oprganization apps.