Ask launchdarkly whether to use Depot as our default builder (#3900)
* ask launchdarkly whether to use Depot as our default builder

i changed the depot flag to be a string, specifically since just saying
that the default is false isn't quite right. like with deploy-retries,
the default value of the flag is going to be set to "auto". In order to
be backwards compatible, we make just using the `--depot` flag be equal
to true

* add some more tracing to the LD client

this will be useful for seeing what flags are enabled for what users

* add some more tracing to depot.go