Support all VMSizeFlags in `fly postgres create` (#4216)
Before this change, `fly postgres create` only supported a set of
bespoke VM sizes via the --vm-size flag, making it impossible to create
a postgres cluster with, say, 1 shared CPU and 1 GB of memory.

As of this change, we create the underlying machine using the entire
`MachineGuest` set of options, which allows for additional
customization.

The legacy prompt and error behavior given no --vm-size or an invalid
--vm-size is preserved.